feat: Fixed the problem of failure to convert schema when right-clicking
