+++ CHANGELOG.md
@@ -1,4 +1,10 @@
+### 0.7.2 (2024-04-24)
+
+1. 修复右键转换`swagger`不生效
+2. 优化将`schema`相关逻辑抽离到`schemaToTs`文件
+
 ### 0.7.1 (2024-04-23)
+
 1. 修复`name`为`undefined`报错
 
 ### 0.7.0 (2024-04-23)
