+++ src/interface/SchemaToTs.interface.ts
@@ -1,0 +1,3 @@
+export interface ISchemaToTs {
+
+}
